+++ .gitignore
@@ -1,2 +1,2 @@
-generate_count_matrix_ADTs
+generate_feature_barcode_matrices
 .DS_Store
