+++ barcode_utils.hpp
@@ -17,11 +17,11 @@
 
 
 struct ValueType {
-	int vid;
+	int32_t vid;
 	uint32_t mask;  // positions of mismatch
 
 	ValueType() : vid(-1), mask(0) {}
-	ValueType(int vid, uint32_t mask) : vid(vid), mask(mask) {}
+	ValueType(int32_t vid, uint32_t mask) : vid(vid), mask(mask) {}
 };
 
 typedef std::unordered_map<uint64_t, ValueType> HashType;
@@ -29,11 +29,11 @@ typedef HashType::iterator HashIterType;
 
 struct IndexType {
 	uint64_t bid;
-	int vid;
+	int32_t vid;
 	uint32_t mask;
 
 	IndexType(): bid(0), vid(-1), mask(0) {}
-	IndexType(uint64_t bid, int vid, uint32_t mask): bid(bid), vid(vid), mask(mask) {}
+	IndexType(uint64_t bid, int32_t vid, uint32_t mask): bid(bid), vid(vid), mask(mask) {}
 };
 
 const int STEP = 3;
@@ -146,38 +146,6 @@ inline void trim(std::string &s) {
     ltrim(s);
 }
 
-inline void group_by_modality(HashType& index_dict, std::vector<std::string>& index_names) {
-	std::vector<int> indices(index_names.size());
-	std::iota(indices.begin(), indices.end(), 0);
-	std::sort(indices.begin(), indices.end(),
-		[&index_names](int l, int r) {
-			std::string s1 = index_names[l];
-			std::string s2 = index_names[r];
-			return s1.substr(s1.find_first_of(',') + 1) < s2.substr(s2.find_first_of(',') + 1);
-		}
-	);
-
-	bool already_sorted = true;
-	for (int i = 0; i < indices.size(); ++i)
-		if (indices[i] != i) {
-			already_sorted = false;
-			break;
-		}
-
-	// No action if barcodes are already grouped by modality column
-	if (already_sorted) return;
-
-	std::vector<int> idx_map(indices.size(), -1);
-	std::vector<std::string> tmp_names(index_names);
-	for (int i = 0; i < indices.size(); ++i) {
-		idx_map[indices[i]] = i;
-		index_names[i] = tmp_names[indices[i]];
-	}
-	for (auto iter = index_dict.begin(); iter != index_dict.end(); ++iter) {
-		iter->second.vid = idx_map[iter->second.vid];
-	}
-}
-
 inline void parse_one_line(const std::string& line, int& n_barcodes, int& barcode_len, HashType& index_dict, std::vector<std::string>& index_names, int max_mismatch, std::queue<IndexType>* buffer) {
 	std::string index_name, index_seq;
 	std::size_t pos;
@@ -202,15 +170,12 @@ inline void parse_one_line(const std::string& line, int& n_barcodes, int& barcod
 	insert(index_dict, barcode_to_binary(index_seq), ValueType(n_barcodes, 0), index_record);
 	buffer->emplace(index_record);
 
-	//if (max_mismatch == 1) mutate_index_one_mismatch(index_dict, index_seq, n_barcodes);
-	//else mutate_index(index_dict, barcode_to_binary(index_seq), index_seq.length(), n_barcodes, max_mismatch, 0, 0);
-
 	index_names.emplace_back(index_name);
 	++n_barcodes;
 }
 
 inline void skip_bom(std::string& line) {
-	size_t start = 0;
+	std::size_t start = 0;
 
 	if (line.length() >= 3 && line.substr(0, 3) == "\xEF\xBB\xBF")   // UTF-8
 		start = 3;
@@ -228,22 +193,20 @@ void insert_index_mutations(HashType& index_dict, int barcode_len, int max_misma
 	while (cur_mismatch <= max_mismatch) {
 		while (!buffer1->empty()) {
 			IndexType& index_val = buffer1->front();
-			if (std::popcount(index_val.mask) < max_mismatch) {
-				// Start from the next position since the last mutation
-				int start_pos = std::numeric_limits<uint32_t>::digits - std::countl_zero(index_val.mask);
-				for (int i = start_pos; i < barcode_len; ++i) {
-					uint64_t val = index_val.bid & aux_arr[i][NNUC];
-					uint32_t mask = index_val.mask | (1 << i);
-					for (int j = 0; j < NNUC; ++j)
-						if (val != aux_arr[i][j]) {
-							uint64_t bid_new = index_val.bid - val + aux_arr[i][j];
-							IndexType index_ret;
-							if (!insert(index_dict, bid_new, ValueType(index_val.vid, mask), index_ret))
-								early_stop = true;
-							if (cur_mismatch < max_mismatch)
-								buffer2->emplace(bid_new, index_val.vid, mask);
-						}
-				}
+			// Start from the next position since the last mutation
+			int start_pos = std::numeric_limits<uint32_t>::digits - std::countl_zero(index_val.mask);
+			for (int i = start_pos; i < barcode_len; ++i) {
+				uint64_t val = index_val.bid & aux_arr[i][NNUC];
+				uint32_t mask = index_val.mask | (1 << i);
+				for (int j = 0; j < NNUC; ++j)
+					if (val != aux_arr[i][j]) {
+						uint64_t bid_new = index_val.bid - val + aux_arr[i][j];
+						IndexType index_ret;
+						if (!insert(index_dict, bid_new, ValueType(index_val.vid, mask), index_ret))
+							early_stop = true;
+						if (cur_mismatch < max_mismatch)
+							buffer2->emplace(bid_new, index_val.vid, mask);
+					}
 			}
 			buffer1->pop();
 		}
@@ -292,12 +255,72 @@ void parse_sample_sheet(const std::string& sample_sheet_file, int& n_barcodes, i
 
 	insert_index_mutations(index_dict, barcode_len, max_mismatch, buffer1, buffer2, verbose);
 
+	delete buffer1;
+	delete buffer2;
+
 	if (verbose) printf("%s is parsed. n_barcodes = %d, and barcode_len = %d.\n", sample_sheet_file.c_str(), n_barcodes, barcode_len);
 
 	int n_amb = 0;
 	for (auto&& kv : index_dict)
 		if (kv.second.vid < 0) ++n_amb;
 	if (verbose) printf("In the index, %d out of %d items are ambigious, percentage = %.2f%%.\n", n_amb, (int)index_dict.size(), n_amb * 100.0 / index_dict.size());
+}
+
+bool parse_feature_names(int n_feature, HashType& feature_index, std::vector<std::string>& feature_names, int& n_cat, std::vector<std::string>& cat_names, std::vector<int>& cat_nfs, std::vector<int>& feature_categories) {
+	std::size_t pos;
+
+	pos = feature_names[0].find_first_of(',');
+	if (pos == std::string::npos) {
+		n_cat = 1;
+		return false;
+	}
+
+	std::vector<std::string> cnames;
+	for (int i = 0; i < n_feature; ++i) {
+		pos = feature_names[i].find_first_of(',');
+		assert(pos != std::string::npos);
+		cnames.push_back(feature_names[i].substr(pos + 1));
+		feature_names[i] = feature_names[i].substr(0, pos);
+	}
+
+	// Group features by modality
+	std::vector<int> indices(feature_names.size());
+	std::iota(indices.begin(), indices.end(), 0);
+	std::stable_sort(indices.begin(), indices.end(),
+		[&cnames](int l, int r) {
+			return cnames[l] < cnames[r];
+		}
+	);
+
+	std::vector<int> idx_map(indices.size(), -1);
+	std::vector<std::string> tmp_fnames(feature_names);
+	std::vector<std::string> tmp_cnames(cnames);
+	for (int i = 0; i < indices.size(); ++i) {
+		idx_map[indices[i]] = i;
+		if (indices[i] != i) {
+			feature_names[i] = tmp_fnames[indices[i]];
+			cnames[i] = tmp_cnames[indices[i]];
+		}
+	}
+	for (auto iter = feature_index.begin(); iter != feature_index.end(); ++iter)
+		iter->second.vid = idx_map[iter->second.vid];
+
+	// Get modality start and end indices
+	n_cat = 0;
+	cat_names.clear();
+	cat_nfs.clear();
+	feature_categories.resize(n_feature, 0);
+	for (int i = 0; i < n_feature; ++i) {
+		if (n_cat == 0 || cat_names.back() != cnames[i]) {
+			cat_names.push_back(cnames[i]);
+			cat_nfs.push_back(i);
+			++n_cat;
+		}
+		feature_categories[i] = n_cat - 1;
+	}
+	cat_nfs.push_back(n_feature);
+
+	return true;
 }
 
 #endif
